替换new Thread为ThreadUtil

+++ app/src/main/java/com/zlm/hp/application/HPApplication.java
@@ -19,6 +19,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import base.utils.PreferencesUtil;
+import base.utils.ThreadUtil;
 
 /**
  * Created by zhangliangming on 2017/7/15.
@@ -174,20 +175,18 @@ public class HPApplication extends Application {
 
     public void setCurAudioInfos(final List<AudioInfo> curAudioInfos) {
         this.curAudioInfos = curAudioInfos;
-        new Thread() {
-            @Override
-            public void run() {
+        ThreadUtil.runInThread(new Runnable() {
+            @Override public void run() {
+
                 String filePath = ResourceFileUtil.getFilePath(getApplicationContext(), ResourceConstants.PATH_CACHE_SERIALIZABLE, "curAudioInfos.ser");
                 if (curAudioInfos != null) {
                     SerializableObjUtil.saveObj(filePath, curAudioInfos);
                 } else {
                     File file = new File(filePath);
-                    if (file.exists()) {
-                        file.delete();
-                    }
+                    if (file.exists()) {  file.delete();  }
                 }
-            }
-        }.start();
+
+            }  });
     }
 
     public AudioInfo getCurAudioInfo() {
@@ -200,21 +199,17 @@ public class HPApplication extends Application {
 
     public void setCurAudioInfo(final AudioInfo curAudioInfo) {
         this.curAudioInfo = curAudioInfo;
-        new Thread() {
-            @Override
-            public void run() {
+        ThreadUtil.runInThread(new Runnable() {
+            @Override public void run() {
+
                 String filePath = ResourceFileUtil.getFilePath(getApplicationContext(), ResourceConstants.PATH_CACHE_SERIALIZABLE, "curAudioInfo.ser");
                 if (curAudioInfo != null) {
                     SerializableObjUtil.saveObj(filePath, curAudioInfo);
                 } else {
                     File file = new File(filePath);
-                    if (file.exists()) {
-                        file.delete();
-                    }
+                    if (file.exists()) { file.delete(); }
                 }
-            }
-        }.start();
-
+            }  });
 
     }
 
@@ -228,20 +223,18 @@ public class HPApplication extends Application {
 
     public void setCurAudioMessage(final AudioMessage curAudioMessage) {
         this.curAudioMessage = curAudioMessage;
-        new Thread() {
-            @Override
-            public void run() {
+        ThreadUtil.runInThread(new Runnable() {
+            @Override public void run() {
+
                 String filePath = ResourceFileUtil.getFilePath(getApplicationContext(), ResourceConstants.PATH_CACHE_SERIALIZABLE, "curAudioMessage.ser");
                 if (curAudioMessage != null) {
                     SerializableObjUtil.saveObj(filePath, curAudioMessage);
                 } else {
                     File file = new File(filePath);
-                    if (file.exists()) {
-                        file.delete();
-                    }
+                    if (file.exists()) {  file.delete(); }
                 }
-            }
-        }.start();
+            } });
+
     }
 
     public RankListResult getRankListResult() {
@@ -254,20 +247,17 @@ public class HPApplication extends Application {
 
     public void setRankListResult(final RankListResult rankListResult) {
         this.rankListResult = rankListResult;
-        new Thread() {
-            @Override
-            public void run() {
+        ThreadUtil.runInThread(new Runnable() {
+            @Override public void run() {
+
                 String filePath = ResourceFileUtil.getFilePath(getApplicationContext(), ResourceConstants.PATH_CACHE_SERIALIZABLE, "rankListResult.ser");
                 if (rankListResult != null) {
                     SerializableObjUtil.saveObj(filePath, rankListResult);
                 } else {
                     File file = new File(filePath);
-                    if (file.exists()) {
-                        file.delete();
-                    }
+                    if (file.exists()) { file.delete();  }
                 }
-            }
-        }.start();
+            }  });
     }
 
     public boolean isLrcSeekTo() {
